Changed the exit button, no page refresh.
Hope you don't mind the change, no reason to refresh the page every time
the popup closes, i used the java script to remove it. Also i changed
the style of positioning so it fits better. It is off of the window
right now, but it can be moved to fit next to the window.